refactor: inline singleflight wrapper into EnsureClone
ensureCloneShared took an fn callback so unit tests could exercise the
singleflight wiring with synthetic operations, but the production
caller passed a closure over ensureCloneLocked. The synthetic-fn tests
were assertions about singleflight.Group.DoChan's behavior — library
contract, not the wrapper's. The seam invited that mode of testing.

Fold the slot logic directly into EnsureClone and rename the inner
HEAD-or-fetch dispatch to ensureCloneNow. Replace the synthetic-fn
tests with one integration test that exercises the canceled-context
pre-check end-to-end (TestEnsureCloneShortCircuitsCanceledContext)
and a unit test for the key shape (TestEnsureCloneKey).

The remaining invariants — detached runner ctx, deadline cap, dedup
of concurrent callers — are covered by the integration tests in
clone_test.go that exercise the real cloneBare/fetch paths, and the
key test pins the only piece of the singleflight contract that's
ours to break.
